[ IS-31 ] Добавить возможность админу назначать модераторов. Запретить админу забанить самого себя

+++ src/main/java/edu/example/service/UserService.java
@@ -2,17 +2,21 @@ package edu.example.service;
 
 import com.querydsl.core.types.Predicate;
 import edu.example.exception.EntityNotFoundException;
+import edu.example.model.Role;
 import edu.example.model.User;
 import edu.example.repository.TokenRepository;
 import edu.example.repository.UserInfoRepository;
 import edu.example.repository.UserRepository;
+import jakarta.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import static java.util.Objects.nonNull;
 
 
 @Service
@@ -39,28 +43,40 @@ public class UserService {
     }
 
     /**
-     * Set user's new ban status. If the user is getting banned, then all their tokens will be deactivated.
+     * Set user's new ban status or role. If the user is getting banned, then all their tokens will be deactivated.
      * This way the user won't be able to perform any authorized requests after ban.
      * @param userId id of a user
      * @param isBanned new ban status
+     * @param role new role
      * @return updated user instance
      */
     @Transactional
-    public User updateUserBanStatus(Long userId, boolean isBanned) {
+    public User updateUser(Long userId, @Nullable Boolean isBanned, @Nullable Role role) {
         var user = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("User not found"));
 
-        if (isBanned) {
-            tokenService.deactivateUserTokens(user.getId());
+        if (user.getRole().equals(Role.ADMIN)) {
+            throw new AccessDeniedException("Updating admin is not allowed");
         }
-        user.setBanned(isBanned);
-        user = userRepository.save(user);
-        return user;
+
+        if (nonNull(isBanned)) {
+            if (isBanned) {
+                tokenService.deactivateUserTokens(user.getId());
+            }
+            user.setBanned(isBanned);
+        }
+        if (nonNull(role)) {
+            user.setRole(role);
+        }
+        return userRepository.save(user);
     }
 
     @Transactional
     public void deleteUser(Long id) {
         var user = userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
+        if (user.getRole().equals(Role.ADMIN)) {
+            throw new AccessDeniedException("Deleting admin account is not allowed");
+        }
         try {
             userInfoRepository.deleteById(user.getUsername());
         } catch (EntityNotFoundException ignored) {}
